requirements.txt
streamlit
pandas
pymongo
plotly
python-dotenv
streamlit-option-menu
numpy
dnspython